+++ cmd/terramate/cli/gitlab/doc.go
@@ -1,0 +1,5 @@
+// Copyright 2024 Terramate GmbH
+// SPDX-License-Identifier: MPL-2.0
+
+// Package gitlab provides a SDK and helpers for the gitlab provider.
+package gitlab